Changed Assignment to also have a string containing its due time (since we combined the time in the date components, we could no longer use date components as a key!  Since each one would have a different time).  Also changed some ints to NSUIntegers.

+++ MyMaret/AnnouncementsTableViewController.m
@@ -105,7 +105,7 @@
         } else {
             // Make an array of all the NSIndexPaths to insert
             NSMutableArray *rowsToInsert = [NSMutableArray array];
-            for (int i = 0; i < numAdded; i++) {
+            for (NSUInteger i = 0; i < numAdded; i++) {
                 NSIndexPath *ip = [NSIndexPath indexPathForRow:i
                                                      inSection:0];
                 [rowsToInsert addObject:ip];
